+++ src/packager/packager/__init__.py
@@ -1,4 +1,5 @@
-from .constants import Constants, KEYED, PseudoContext  # noqa: F401
+from .types import KEYED, PseudoContext  # noqa: F401
+from .constants import Constants  # noqa: F401
 from .index import handler  # noqa: F401
 from .handler import Handler  # noqa: F401
 from .ssm_parameter_store import SSMParameterStore  # noqa: F401
